+++ doc/release_notes.rst
@@ -3,7 +3,8 @@ Release notes
 
 * Initial implementation of the Cyg-X1 config/example.
 * First implementation of a script for the calculation of the MDP.
-
+* Significant refactoring of the srcmodel/spectrum.py module.
+  
 
 *ximpol (0.39.4) - Thu, 07 Apr 2016 13:59:05 +0200*